fix: externalized code and removed code repetition from store
