+++ src/store.js
@@ -1,31 +1,8 @@
 import Vue from "vue";
 import Vuex from "vuex";
+import { groupRepeatedUnits, createRepetition } from "./utils/repetition";
 
 Vue.use(Vuex);
-const groupRepeatedUnits = (templateArray = ["1fr"]) => {
-  const groups = [[templateArray.shift()]];
-  for (const templateUnit of templateArray) {
-    const lastGroup = groups[groups.length - 1];
-    if (lastGroup.indexOf(templateUnit) !== -1) {
-      lastGroup.push(templateUnit);
-    } else {
-      groups.push([templateUnit]);
-    }
-  }
-  return groups;
-};
-
-const createRepetition = (groups, maxRepetition = 1) => {
-  return groups
-    .map(group =>
-      // If you want to add repetition only when a measure is repeated more than 1 time,
-      // change maxRepetition value
-      group.length === maxRepetition
-        ? group.join(" ")
-        : `repeat(${group.length}, ${group[0]})`
-    )
-    .join(" ");
-};
 
 export default new Vuex.Store({
   state: {
@@ -39,15 +16,11 @@ export default new Vuex.Store({
   },
   getters: {
     colTemplate(state) {
-      // Maybe this mapping can be executed inside groupRepeatedUnits
-      // to avoid code repetition
-      const templateArray = state.colArr.map(i => i["unit"]);
-      const unitGroups = groupRepeatedUnits(templateArray);
+      const unitGroups = groupRepeatedUnits(state.colArr);
       return createRepetition(unitGroups);
     },
     rowTemplate(state) {
-      const templateArray = state.rowArr.map(i => i["unit"]);
-      const unitGroups = groupRepeatedUnits(templateArray);
+      const unitGroups = groupRepeatedUnits(state.rowArr);
       return createRepetition(unitGroups);
     },
     divNum(state) {
